test(cloud): isolate inference admission mocks
